Move many classes to more sensible locations
