+++ src/main/java/basemod/console/ConsoleInputProcessor.java
@@ -1,6 +1,9 @@
-package basemod;
+package basemod.console;
 
 import com.badlogic.gdx.Input.Keys;
+
+import basemod.DevConsole;
+
 import com.badlogic.gdx.InputProcessor;
 
 import java.awt.event.KeyEvent;
