+++ src/main/java/basemod/init/EditCharactersInit.java
@@ -1,5 +1,6 @@
-package basemod;
+package basemod.init;
 
+import basemod.BaseMod;
 import basemod.interfaces.PostInitializeSubscriber;
 
 public class EditCharactersInit implements PostInitializeSubscriber {
